Sync code to the newest

+++ src/drivers/Qt/sdl-video.h
@@ -6,7 +6,7 @@
 #include <SDL.h>
 #endif
 
-uint32 PtoV(uint16 x, uint16 y);
+uint32 PtoV(double x, double y);
 bool FCEUD_ShouldDrawInputAids();
 bool FCEUI_AviDisableMovieMessages();
 bool FCEUI_AviEnableHUDrecording();
